Restructure verifiable-db query + fix dependent crates
