+++ groth16-framework/tests/common/context.rs
@@ -3,14 +3,13 @@
 use super::{NUM_PREPROCESSING_IO, NUM_QUERY_IO};
 use groth16_framework::{compile_and_generate_assets, utils::clone_circuit_data};
 use mp2_common::{C, D, F};
+use mp2_test::circuit::TestDummyCircuit;
 use recursion_framework::framework_testing::TestingRecursiveCircuits;
 use verifiable_db::{
-    api::WrapCircuitParams,
-    revelation::api::Parameters as RevelationParameters,
-    test_utils::{
+    api::WrapCircuitParams, query::pi_len, revelation::api::Parameters as RevelationParameters, test_utils::{
         INDEX_TREE_MAX_DEPTH, MAX_NUM_COLUMNS, MAX_NUM_ITEMS_PER_OUTPUT, MAX_NUM_OUTPUTS,
         MAX_NUM_PLACEHOLDERS, MAX_NUM_PREDICATE_OPS, MAX_NUM_RESULT_OPS, ROW_TREE_MAX_DEPTH,
-    },
+    }
 };
 
 /// Test context
@@ -40,7 +39,8 @@ impl TestContext {
 
         // Generate a fake query circuit set.
         let query_circuits = TestingRecursiveCircuits::<F, C, D, NUM_QUERY_IO>::default();
-
+        let dummy_universal_circuit = TestDummyCircuit::<{pi_len::<MAX_NUM_ITEMS_PER_OUTPUT>()}>::build();
+        
         // Create the revelation parameters.
         let revelation_params = RevelationParameters::<
             ROW_TREE_MAX_DEPTH,
@@ -53,7 +53,7 @@ impl TestContext {
             MAX_NUM_PLACEHOLDERS,
         >::build(
             query_circuits.get_recursive_circuit_set(), // unused, so we provide a dummy one
-            query_circuits.get_recursive_circuit_set(),
+            dummy_universal_circuit.circuit_data().verifier_data(),
             preprocessing_circuits.get_recursive_circuit_set(),
             preprocessing_circuits
                 .verifier_data_for_input_proofs::<1>()
