+++ mp2-v1/tests/common/cases/mod.rs
@@ -11,7 +11,6 @@ use super::table::Table;
 
 pub mod contract;
 pub mod indexing;
-pub mod planner;
 pub mod query;
 pub mod table_source;
