+++ mp2-v1/tests/common/cases/query/mod.rs
@@ -11,7 +11,7 @@ use log::info;
 use mp2_v1::{
     api::MetadataHash, indexing::block::BlockPrimaryIndex, query::planner::execute_row_query,
 };
-use parsil::{parse_and_validate, utils::ParsilSettingsBuilder, PlaceholderSettings};
+use parsil::{assembler::DynamicCircuitPis, parse_and_validate, utils::ParsilSettingsBuilder, ParsilSettings, PlaceholderSettings};
 use simple_select_queries::{
     cook_query_no_matching_rows, cook_query_too_big_offset, cook_query_with_distinct,
     cook_query_with_matching_rows, cook_query_with_max_num_matching_rows,
@@ -23,7 +23,7 @@ use verifiable_db::query::{
     computational_hash_ids::Output, universal_circuit::universal_circuit_inputs::Placeholders,
 };
 
-use crate::common::{cases::planner::QueryPlanner, table::Table, TableInfo, TestContext};
+use crate::common::{table::{Table, TableColumns}, TableInfo, TestContext};
 
 use super::table_source::TableSource;
 
@@ -55,14 +55,6 @@ pub type GlobalCircuitInput = verifiable_db::api::QueryCircuitInput<
 >;
 
 pub type QueryCircuitInput = verifiable_db::query::api::CircuitInput<
-    MAX_NUM_COLUMNS,
-    MAX_NUM_PREDICATE_OPS,
-    MAX_NUM_RESULT_OPS,
-    MAX_NUM_ITEMS_PER_OUTPUT,
->;
-
-#[cfg(feature = "batching_circuits")]
-pub type BatchingQueryCircuitInput = verifiable_db::query::batching::CircuitInput<
     NUM_CHUNKS,
     NUM_ROWS,
     ROW_TREE_MAX_DEPTH,
@@ -92,6 +84,17 @@ pub struct QueryCooking {
     pub(crate) max_block: BlockPrimaryIndex,
     pub(crate) limit: Option<u32>,
     pub(crate) offset: Option<u32>,
+}
+
+pub(crate) struct QueryPlanner<'a> {
+    pub(crate) query: QueryCooking,
+    pub(crate) pis: &'a DynamicCircuitPis,
+    pub(crate) ctx: &'a mut TestContext,
+    pub(crate) settings: &'a ParsilSettings<&'a Table>,
+    // useful for non existence since we need to search in both trees the places to prove
+    // the fact a given node doesn't exist
+    pub(crate) table: &'a Table,
+    pub(crate) columns: TableColumns,
 }
 
 pub async fn test_query(ctx: &mut TestContext, table: Table, t: TableInfo) -> Result<()> {
@@ -215,14 +218,10 @@ async fn test_query_mapping(
     match pis.result.query_variant() {
         Output::Aggregation => {
             prove_aggregation_query(
-                ctx,
-                table,
-                query_info,
                 parsed,
-                &settings,
                 res,
                 *table_hash,
-                pis,
+                &mut planner,
             )
             .await
         }
