+++ parsil/src/assembler.rs
@@ -15,7 +15,7 @@ use sqlparser::ast::{
     SelectItem, SetExpr, TableAlias, TableFactor, UnaryOperator, Value,
 };
 use verifiable_db::query::{
-    aggregation::{QueryBoundSource, QueryBounds},
+    utils::{QueryBoundSource, QueryBounds},
     computational_hash_ids::{AggregationOperation, Operation, PlaceholderIdentifier},
     universal_circuit::universal_circuit_inputs::{
         BasicOperation, InputOperand, OutputItem, Placeholders, ResultStructure,
