--- verifiable-db/src/query/batching/mod.rs
@@ -1,2 +1,0 @@
-pub(crate) mod circuits;
-pub(crate) mod row_chunk;
